:recycle: (global) Converts everything to microservices
This is the big scary refactor that turns all of Aibex into a collection of microservices. They
include bleat (api), capra (www), dev-convo, dev-nginx, dev-postgres, dev-rabbit, judas (jobs), and
karl (chat). The idea is to make capra as lean as possible, relying on API and Chat endpoints for
other parts of aibex. This not only decouples parts of our infra, it will make bug fixing, testing,
and releasing easier. To support this set of changes, every microservice required its own ai.yaml,
parts of Capra were carved away, and each piece was tested alone. As of this commit, Judas /
dev-rabbit are still in experimental stage, and dev-convo is retired.


Former-commit-id: 3e545daf6394deae160a2f773df9a164c02e8fb9